Create Task_3.c
Check, whether the number 6321246989963 is prime or not. If not, find
its factorization (represent as a product of prime numbers).
Hint: this is a big number, which does not fit into 32 bits (a normal int
type). But in MS Visual C++, you may use long long int to store
very big integers and "%I64d" to print them. You may use a suffix "LL"
to ensure, that the value is treated as a very long integer (in Visual C++ it
works even without LL):
 long long int n;
 n = 6321246989963LL;
Warning: Depending on the implementation, program may take long time
to yield the results. Firstly check if number 6321246989963 is a prime,
then test code for factorization finding, for smaller number (e.g. 123123).