Instruction
Write a program copying a text file from stdin to stdout in such a way,
that each line of text is written twice. Use ONLY functions reading and
writing single characters and an array storing a single line of text. You
may assume, that the length of a text line does not exceed 100 characters.
It is recommended to use getchar(…) and putchar(…)