Instruction
Write a function finding the greatest common divisor (GCD) of two
integers, using the Euclid's algorithm.
The GCD of two integers is equal to one of them if these number are
equal. If they are not, then the GCD(x,y) is equal to GCD(x-y,y) when
x>y or GCD(y-x,x) when y>x.