Instruction
3. Read about lists. Ask the user for 5 numbers. Store them in a singly linked list. Then print the list.
4. Write a function inserting an element between existing elements of a list.
The function should accept the address of the element (or index if such
functionality was implemented), after which a new element should be
inserted (it will not be possible to add an element before the first element
of the list). Verify how the function works by printing all elements of the
modified list.